feat: add `CompositeActionMatcher`
